Add WorkVisaApplicationComment resource and admin registration
Introduced the WorkVisaApplicationCommentResource for managing comments related to work visa applications. Registered the WorkVisaApplicationComment model in the Django admin with ImportExport capabilities, enhancing the administrative interface for managing application comments. Updated the WorkVisaApplicationComment model to improve the display logic for sender information, ensuring clarity in communication between applicants and admins.